:computer: S3 direct uploader
