+++ package.json
@@ -13,7 +13,6 @@
     "@braid/vue-formulate": "^2.5.2",
     "axios": "^0.21.1",
     "core-js": "^3.13.0",
-    "s3-vuetify-file-browser": "^1.1.7",
     "vue": "^2.6.11",
     "vue-router": "^3.4.9",
     "vuetify": "^2.4.10",
